requirements.txt
requests>=2.28
SQLAlchemy>=1.4
pandas>=2.0
streamlit>=1.20
python-dotenv>=1.0
plotly>=5.0
tqdm>=4.65
